requirements.txt
streamlit==1.10.0
transformers==4.11.3
torch==1.9.0
Pillow==8.3.2
